functioning reformatted
